refactor: consolidate top-level scripts into src/ and scripts/
Move the four loose Python modules out of the repo root:

- server.py        -> src/server.py
- classify.py      -> src/classifier.py
- main.py          -> scripts/check_mlflow.py
- app.py           -> folded into src/api/routes.py (deleted)

The old app.py was a misnamed router (not a FastAPI app) and shared a
circular dependency with server.py: server.py imported its router, and
the router's handlers reached back for the model pipeline via deferred
function-local imports plus a `# noqa: E402` on the forward import.

Routing is now consolidated in src/api/routes.py and the pipeline is
stored on app.state.pipeline (set in lifespan, read via Request), which
removes both the cycle and the workarounds that came with it.

Also updates Dockerfile (COPY src/ only; CMD uvicorn src.server:app),
README run commands and project structure tree, conftest fixture, and
adds a CHANGELOG entry.
