+++ Dockerfile
@@ -13,7 +13,6 @@ RUN useradd --create-home --shell /usr/sbin/nologin appuser
 WORKDIR /app
 COPY pyproject.toml ./
 RUN pip install --no-cache-dir .
-COPY app.py classify.py server.py ./
 COPY src/ src/
 COPY models/ models/
 COPY --from=frontend-build /app/frontend/dist frontend/dist
@@ -21,4 +20,4 @@ RUN chown -R appuser:appuser /app
 USER appuser
 
 EXPOSE 8080
-CMD ["uvicorn", "server:app", "--host", "0.0.0.0", "--port", "8080"]
+CMD ["uvicorn", "src.server:app", "--host", "0.0.0.0", "--port", "8080"]
