+++ src/api/routes.py
@@ -1,12 +1,55 @@
 import json
 from json import JSONDecodeError
 
-from fastapi import APIRouter, HTTPException
+import fitz
+from fastapi import APIRouter, HTTPException, Request, UploadFile
 
-from .database import get_classification, get_history, get_stats
-from .models import HistoryEntry, HistoryResponse, StatsResponse
+from src.classifier import extract_text_from_bytes, predict_from_text
+
+from .database import get_classification, get_history, get_stats, save_classification
+from .models import HistoryEntry, HistoryResponse, PredictionResponse, StatsResponse
 
 router = APIRouter()
+
+
+@router.get("/health")
+def health(request: Request):
+    pipeline = getattr(request.app.state, "pipeline", None)
+    return {"status": "ok", "model_loaded": pipeline is not None}
+
+
+@router.post("/predict", response_model=PredictionResponse)
+async def predict_pdf(request: Request, file: UploadFile):
+    pipeline = getattr(request.app.state, "pipeline", None)
+    if pipeline is None:
+        raise HTTPException(
+            status_code=503, detail="Model not loaded. Train a model first."
+        )
+
+    pdf_bytes = await file.read()
+    try:
+        text = extract_text_from_bytes(pdf_bytes)
+    except fitz.FileDataError as exc:
+        raise HTTPException(
+            status_code=422, detail="File is not a readable PDF"
+        ) from exc
+
+    if not text.strip():
+        raise HTTPException(status_code=422, detail="PDF valid but not processable")
+
+    result = predict_from_text(pipeline, text)
+
+    confidence = max(result["probabilities"].values())
+    await save_classification(
+        filename=file.filename or "unknown.pdf",
+        label=result["label"],
+        confidence=confidence,
+        probabilities=result["probabilities"],
+        text_preview=text[:500] if text else None,
+        file_size=len(pdf_bytes),
+    )
+
+    return result
 
 
 @router.get("/history", response_model=HistoryResponse)
