+++ src/server.py
@@ -4,32 +4,22 @@ from contextlib import asynccontextmanager
 from fastapi import FastAPI
 from fastapi.staticfiles import StaticFiles
 
-from classify import load_model
 from src.api.database import init_db
-
-_pipeline = None
-
-
-def get_pipeline():
-    return _pipeline
+from src.api.routes import router as api_router
+from src.classifier import load_model
 
 
 @asynccontextmanager
 async def lifespan(app: FastAPI):
-    global _pipeline
-    _pipeline = load_model()
+    app.state.pipeline = load_model()
     await init_db()
     yield
 
 
 app = FastAPI(title="DocQFlow", lifespan=lifespan)
-
-from app import router as classify_router  # noqa: E402
-from src.api.routes import router as api_router  # noqa: E402
-
-app.include_router(classify_router, prefix="/api")
 app.include_router(api_router, prefix="/api")
 
-dist_dir = os.path.join(os.path.dirname(__file__), "frontend", "dist")
+REPO_ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+dist_dir = os.path.join(REPO_ROOT, "frontend", "dist")
 if os.path.isdir(dist_dir):
     app.mount("/", StaticFiles(directory=dist_dir, html=True), name="static")
